Switched to a float delta system

+++ src/com/osreboot/ridhvl/menu/HvlButton.java
@@ -37,7 +37,7 @@ public abstract class HvlButton {
 
 	public void onPressing(int buttonArg){}
 	public void onTriggered(){}
-	public void draw(long delta){}
+	public void draw(float delta){}
 
 	public final boolean isBeingPressed(int buttonArg){//TODO account for HvlDisplayMode
 		return Mouse.isInsideWindow() && Mouse.isButtonDown(buttonArg) && Mouse.getX() > xLocation && inversionHeight - Mouse.getY() > yLocation && Mouse.getX() < xLocation + xLength && inversionHeight - Mouse.getY() < yLocation + yLength;
@@ -47,7 +47,7 @@ public abstract class HvlButton {
 		return Mouse.isInsideWindow() && Mouse.getX() > xLocation && inversionHeight - Mouse.getY() > yLocation && Mouse.getX() < xLocation + xLength && inversionHeight - Mouse.getY() < yLocation + yLength;
 	}
 
-	public final void update(long delta){
+	public final void update(float delta){
 		draw(delta);
 		if(isBeingPressed(0)){
 			onPressing(0);
